# Documentazione/Progettazione/openapi.yaml
openapi: 3.0.3
info:
  title: Monopoly IoT API
  version: 1.0.0
  description: |
    API REST esposte dall'API Gateway del sistema Monopoly IoT.

    Il sistema è composto da 5 microservizi:
    - **api-gateway** (porta 8080) — punto di ingresso unico
    - **user-service** (porta 8081) — gestione utenti
    - **match-service** (porta 8082) — partite, tornei, motore di gioco
    - **board-service** (porta 8083) — sedi, tavoli, zone RFID
    - **iot-gateway** (porta 8085) — acquisizione eventi MQTT, buffer offline

    Tutte le richieste vengono inoltrate dal Gateway ai servizi interni.
servers:
  - url: http://localhost:8080
    description: API Gateway locale
...
paths:
...
  /api/tournaments/create:
    post:
      tags: [Tournaments]
      summary: Crea torneo e genera il primo round
      description: |
        Crea il torneo, registra i giocatori (con shuffle casuale per il bracket),
        e genera i match del primo round con al massimo 4 giocatori per partita.
      operationId: createTournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, playerIds]
              properties:
                name:
                  type: string
                  example: "Torneo Primavera 2026"
                playerIds:
                  type: array
                  minItems: 2
                  items:
                    type: integer
                  example: [1, 2, 3, 4, 5, 6, 7, 8]
                venueId:
                  type: integer
                  format: int64
                  example: 1
      responses:
        "200":
          description: Torneo creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "400":
          description: Numero di giocatori non valido (minimo 2)
  /api/tournaments/{id}:
    delete:
      tags: [Tournaments]
      summary: Elimina torneo e tutti i dati associati (match, giocatori)
      operationId: deleteTournament
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Torneo eliminato
        "404":
          description: Torneo non trovato
  /api/tournaments/{id}/bracket:
    get:
      tags: [Tournaments]
      summary: Tabellone del torneo
      description: Restituisce tutti i match del torneo con i giocatori e il vincitore.
      operationId: getTournamentBracket
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Bracket torneo
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    status:
                      type: string
                    winnerId:
                      type: integer
                      format: int64
                      nullable: true
                    players:
                      type: array
                      items:
                        type: object
                        properties:
                          userId:
                            type: integer
                          balance:
                            type: integer
  /api/tournaments/{id}/advance-round:
    post:
      tags: [Tournaments]
      summary: Genera il turno successivo con i vincitori del round corrente
      description: |
        Raccoglie i vincitori dei match FINISHED, verifica che non siano già stati
        avanzati in turni precedenti, e genera nuovi match (max 4 per partita).
        Richiede almeno 2 vincitori disponibili.
      operationId: advanceRound
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Turno generato
          content:
            application/json:
              schema:
                type: object
                properties:
                  matchesCreated:
                    type: integer
                    example: 2
        "400":
          description: Vincitori insufficienti per avanzare
        "404":
          description: Torneo non trovato

  # ─────────────────────────────────────────────────────────────
  # ANALYTICS (→ match-service :8082)
  # ─────────────────────────────────────────────────────────────
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Identificatore univoco della risorsa
  schemas:
...
    Tournament:
      type: object
      description: Torneo che raggruppa più partite con bracket eliminazione
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          example: "Torneo Primavera 2026"
        gameTypeId:
          type: integer
          format: int64
          default: 1
        status:
          type: string
          enum: [ONGOING, FINISHED, CANCELLED]
          example: "ONGOING"
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        createdByUserId:
          type: integer
          format: int64
        createdAt:
          type: string
...